Fixing namespace issues when using module from pip
When installing ec2utils from pip none of the tools are working
because of invalid imports. The way the code is packed on pip
does not work with the way the modules are imported in the
code. This patch makes ec2utils to work when installed from
pip. It also fixes the use of print to be compatible with
python2 vs. python3