Implement AWS clean up when aborting an upload
When an upload gets aborted with Ctrl + C or SIGTERM / SIGINT
ec2uploadimg left resources like helper instances left in AWS.
This commit implements proper clean up of these AWS resources.

If the upload already started, we finish the current step and run the clean up afterwards.
If the upload did not start but the setup process, we finish the setup and do not start the upload. Finally we clean up the setup (VPC etc).
If neither setup nor upload started, we will just exit.

If the script gets aborted a second time, we just exit without a clean up like we did before.

Implements #194